+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- Reload pipeline step status on patient view (@mkasztelnik)
 
 ### Changed
 - Labels for pipelines list view improved (@mkasztelnik)